+++ packages/dropdown/src/vars/index.ts
@@ -1,0 +1,3 @@
+import { keyMirror } from '@pluralsight/ps-design-system-util'
+export const appearances = keyMirror('default', 'subtle')
+export const sizes = keyMirror('medium', 'small')