+++ packages/position/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@4.1.20...@pluralsight/ps-design-system-position@4.1.21) (2020-10-08)
+
+
+### Bug Fixes
+
+* **position:** update on ref change ([2f76a16](https://github.com/pluralsight/design-system/commit/2f76a16c549dae803ecd613c334202394b1bb059))
+
+
+
+
+
+## [4.1.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@4.1.19...@pluralsight/ps-design-system-position@4.1.20) (2020-10-08)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-position
+
+
+
+
+
 ## [4.1.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@4.1.18...@pluralsight/ps-design-system-position@4.1.19) (2020-10-08)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-position